Adds documentation and handler callback functions
This was the work of a MOB programming session and some cleanup. I left
the commit messages that had useful information. This is our attempt at
allowing the handler to react in different ways to applications shutting
down.

Amos King @adkron <amos@binarynoggin.com>
Mike Schmidt
Justin Schneck

Intention revealing name for apps function

Removes `spawn` that is no longer needed

The spawn was causing a "hacky" callback that we can eliminate since we
moved our failure timer to the `:timer` process. Since the process is
not the same as the one we are in we no longer need to put out handler
in another process. If our handler misbehaves and locks up the
`ApplicationController` the `:timer` will still halt Erlang.

I made a first pass at documenting the handler too.

Amos King @adkron <amos@binarynoggin.com>