Remove support for running MFA's on startup
As part of the change to load applications via boot scripts, support for
running MFA's needs to be removed. The fix is to move the code called by
the MFA into an Application.start callback.

In addition to MFA's not being supported via the boot script, there was
a subtle issue with them that could be confusing: It looks like the MFA
will be called before the applications listed after it have been
started. This isn't guaranteed to be true. Normally, it works since the
`:init` list usually only has a few items and they're not
interdependent. However, when it doesn't hold true, it's confusing since
it really looks like it should.

In hindsight, the use of MFA's looks like it has turned into a
convenience for avoiding the creation of yet another OTP application to
do something simple like report status or start Erlang distribution. It
feels like there are other, more robust ways of accomplishing this..