Remove macro in favor of behaviour for Shoehorn.Handler
This is a breaking change that requires users to replace references to
`use Shoehorn.Handler` with `@behaviour Shoehorn.Handler` and implement
any missing callbacks.

The reason for this change is catch callback mistakes at compile-time.
Previously, the automatically provided functions could be used and cause
errors to not be handled.